Derive parameter types/registers from C for the target function
